Move more BufferLine dependencies into core
