+++ src/Buffer.ts
@@ -7,15 +7,10 @@ import { CircularList, IInsertEvent } from './common/CircularList';
 import { ITerminal, IBuffer, BufferIndex, IBufferStringIterator, IBufferStringIteratorResult } from './Types';
 import { IBufferLine, ICellData, IAttributeData } from './core/Types';
 import { IMarker } from 'xterm';
-import { BufferLine, CellData, AttributeData, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE, WHITESPACE_CELL_CHAR, WHITESPACE_CELL_WIDTH, WHITESPACE_CELL_CODE, CHAR_DATA_WIDTH_INDEX, CHAR_DATA_CHAR_INDEX } from './BufferLine';
+import { BufferLine, CellData, NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE, WHITESPACE_CELL_CHAR, WHITESPACE_CELL_WIDTH, WHITESPACE_CELL_CODE, CHAR_DATA_WIDTH_INDEX, CHAR_DATA_CHAR_INDEX, DEFAULT_ATTR_DATA } from './BufferLine';
 import { reflowLargerApplyNewLayout, reflowLargerCreateNewLayout, reflowLargerGetLinesToRemove, reflowSmallerGetNewLineLengths, getWrappedLineTrimmedLength } from './BufferReflow';
-import { DEFAULT_COLOR } from './renderer/atlas/Types';
 import { EventEmitter2, IEvent } from './common/EventEmitter2';
 import { Disposable } from '../lib/common/Lifecycle';
-
-export const DEFAULT_ATTR = (0 << 18) | (DEFAULT_COLOR << 9) | (256 << 0);
-
-export const DEFAULT_ATTR_DATA = new AttributeData();
 
 export const MAX_BUFFER_SIZE = 4294967295; // 2^32 - 1
