+++ src/renderer/atlas/CharAtlasUtils.ts
@@ -5,7 +5,8 @@
 
 import { ITerminal } from '../../Types';
 import { IColorSet } from '../Types';
-import { DEFAULT_COLOR, ICharAtlasConfig } from './Types';
+import { ICharAtlasConfig } from './Types';
+import { DEFAULT_COLOR } from '../../common/Types';
 
 export function generateConfig(scaledCharWidth: number, scaledCharHeight: number, terminal: ITerminal, colors: IColorSet): ICharAtlasConfig {
   // null out some fields that don't matter
