+++ src/renderer/atlas/Types.ts
@@ -6,7 +6,6 @@
 import { FontWeight } from 'xterm';
 import { IColorSet } from '../Types';
 
-export const DEFAULT_COLOR = 256;
 export const INVERTED_DEFAULT_COLOR = 257;
 export const DIM_OPACITY = 0.5;
